Move FileRepository interface/implementations to Common
