+++ basyx.common/basyx.filerepository-backend-inmemory/pom.xml
@@ -2,34 +2,24 @@
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
-
 	<parent>
 		<groupId>org.eclipse.digitaltwin.basyx</groupId>
-		<artifactId>basyx.submodelrepository</artifactId>
+		<artifactId>basyx.common</artifactId>
 		<version>${revision}</version>
 	</parent>
-
-	<artifactId>basyx.submodelrepository-backend</artifactId>
-
-	<name>BaSyx Submodelrepository backend</name>
-	<description>BaSyx Submodelrepository backend</description>
-
+	<artifactId>basyx.filerepository-backend-inmemory</artifactId>
 	<dependencies>
 		<dependency>
 			<groupId>org.eclipse.digitaltwin.basyx</groupId>
-			<artifactId>basyx.submodelrepository-core</artifactId>
-		</dependency>
-		<dependency>
-		    <groupId>org.springframework.data</groupId>
-		    <artifactId>spring-data-commons</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>commons-io</groupId>
-			<artifactId>commons-io</artifactId>
+			<artifactId>basyx.core</artifactId>
 		</dependency>
 		<dependency>
 			<groupId>org.eclipse.digitaltwin.basyx</groupId>
 			<artifactId>basyx.filerepository-backend</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>commons-io</groupId>
+			<artifactId>commons-io</artifactId>
 		</dependency>
 	</dependencies>
 </project>
