+++ basyx.common/basyx.filerepository-backend/pom.xml
@@ -1,0 +1,16 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>org.eclipse.digitaltwin.basyx</groupId>
+    <artifactId>basyx.common</artifactId>
+    <version>${revision}</version>
+  </parent>
+  <artifactId>basyx.filerepository-backend</artifactId>
+	<name>BaSyx File Repository Backend</name>
+	<dependencies>
+		<dependency>
+			<groupId>org.eclipse.digitaltwin.basyx</groupId>
+			<artifactId>basyx.core</artifactId>
+		</dependency>
+	</dependencies>
+</project>
