+++ basyx.submodelrepository/basyx.submodelrepository-backend-inmemory/pom.xml
@@ -55,5 +55,11 @@
 			<groupId>commons-io</groupId>
 			<artifactId>commons-io</artifactId>
 		</dependency>
+		<dependency>
+			<groupId>org.eclipse.digitaltwin.basyx</groupId>
+			<artifactId>
+				basyx.filerepository-backend-inmemory
+			</artifactId>
+		</dependency>
 	</dependencies>
 </project>
