+++ basyx.submodelrepository/basyx.submodelrepository-backend-mongodb/pom.xml
@@ -48,5 +48,11 @@
 			<groupId>org.apache.tika</groupId>
 			<artifactId>tika-core</artifactId>
 		</dependency>
+		<dependency>
+			<groupId>org.eclipse.digitaltwin.basyx</groupId>
+			<artifactId>
+				basyx.filerepository-backend-mongodb
+			</artifactId>
+		</dependency>
 	</dependencies>
 </project>
